Create CD, Controllers and Docker

+++ src/main/java/com/bastiansmn/scoreserver/service/ScoreService.java
@@ -1,5 +1,6 @@
 package com.bastiansmn.scoreserver.service;
 
+import com.bastiansmn.scoreserver.domain.Namespace;
 import com.bastiansmn.scoreserver.domain.Score;
 import com.bastiansmn.scoreserver.exception.FunctionalException;
 import com.bastiansmn.scoreserver.exception.FunctionalRule;
@@ -10,6 +11,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,7 +21,9 @@ public class ScoreService {
     private final NamespaceRepository namespaceRepository;
 
     public Score create(String namespace, String user_uuid, Long score) throws FunctionalException {
-        if (!this.namespaceRepository.existsByName(namespace))
+        Optional<Namespace> namespaceOptional = namespaceRepository.findByName(namespace);
+
+        if (namespaceOptional.isEmpty())
             throw new FunctionalException(
                     FunctionalRule.NS_NOT_FOUND.getMessage(),
                     HttpStatus.NOT_FOUND
@@ -28,30 +32,51 @@ public class ScoreService {
         Score scoreEntity = Score.builder()
                 .userUUID(user_uuid)
                 .score(score)
+                .namespace(namespaceOptional.get())
                 .build();
 
         return scoreRepository.save(scoreEntity);
     }
 
-    public Score getLastOfUser(String user_uuid) throws FunctionalException {
+    public Score getLastOfUser(String user_uuid, String namespace) throws FunctionalException {
         if (!this.scoreRepository.existsByUserUUID(user_uuid))
             throw new FunctionalException(
                     FunctionalRule.USER_NOT_FOUND.getMessage(), HttpStatus.NOT_FOUND
             );
 
-        return this.scoreRepository.findFirstByUserUUIDOrderByCreatedAtDesc(user_uuid);
+        Optional<Namespace> namespaceOptional = namespaceRepository.findByName(namespace);
+
+        if (namespaceOptional.isEmpty())
+            throw new FunctionalException(
+                    FunctionalRule.NS_NOT_FOUND.getMessage(),
+                    HttpStatus.NOT_FOUND
+            );
+
+        return this.scoreRepository.findFirstByUserUUIDAndNamespaceOrderByCreatedAtDesc(user_uuid,namespaceOptional.get());
     }
 
-    public List<Score> getAllOfUser(String user_uuid) throws FunctionalException {
+    public List<Score> getAllOfUser(String user_uuid, String namespace) throws FunctionalException {
         if (!this.scoreRepository.existsByUserUUID(user_uuid))
             throw new FunctionalException(
                     FunctionalRule.USER_NOT_FOUND.getMessage(), HttpStatus.NOT_FOUND
             );
 
-        return this.scoreRepository.findAllByUserUUIDOrderByCreatedAtAsc(user_uuid);
+        Optional<Namespace> namespaceOptional = namespaceRepository.findByName(namespace);
+
+        if (namespaceOptional.isEmpty())
+            throw new FunctionalException(
+                    FunctionalRule.NS_NOT_FOUND.getMessage(),
+                    HttpStatus.NOT_FOUND
+            );
+
+        return this.scoreRepository.findAllByUserUUIDAndNamespaceOrderByCreatedAtAsc(user_uuid, namespaceOptional.get());
     }
 
-    public void delete(Long score_id) {
+    public void delete(Long score_id) throws FunctionalException {
+        if (!this.scoreRepository.existsById(score_id))
+            throw new FunctionalException(
+                    FunctionalRule.SCORE_NOT_FOUND.getMessage(), HttpStatus.NOT_FOUND
+            );
         this.scoreRepository.deleteById(score_id);
     }
